Created test class for classifications
Added separate file for classification as well as getting accuracy within a minute difference in the delay prediction.
